voice-future-assistant/requirements-prod.txt
google-cloud-aiplatform>=1.36.0
protobuf>=4.23.4
boto3>=1.28.0
wave>=0.0.2
aiohttp
aiortc
elevenlabs
fastapi
uvicorn
numpy
dotenv
google-cloud-texttospeech
google-cloud-speech
google-genai
webrtcvad
amazon-transcribe
azure-cognitiveservices-speech
azure-identity
openai
mutagen
hf_xet
hf_transfer
faster-whisper
edge-tts
google.cloud-storage